[FEAT/#28] 게시글 생성 API

+++ src/main/java/edu/aivle/heatguard_ai_back/repository/NoticeFileRepository.java
@@ -3,5 +3,5 @@ package edu.aivle.heatguard_ai_back.repository;
 import edu.aivle.heatguard_ai_back.entity.NoticeFileEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NoticeFileRepository extends JpaRepository<NoticeFileEntity, Long> {
+public interface NoticeFileRepository extends JpaRepository<NoticeFileEntity, Integer> {
 }
